trying to re write the emit behavior idea in a more simple way

+++ Assets/_Project/Packages/Grain Synth/Scripts/DOTS/Authoring/GrainEmitterAuthoring.cs
@@ -88,7 +88,7 @@ public class GrainEmissionProps
         }
         set
         {
-            _Volume = (int)Mathf.Clamp(value, 0f, 3f);
+            _Volume = Mathf.Clamp(value, 0f, 3f);
         }
     }
